Remove the exposed_modules.ml indirection, expose modules in coccilib
The fix of commit c7cf750d1c44fe28535306628f34d7133611a948 weakens the
type signature of the Exposed_modules module, in a way that restores
the .cmi behavior of older OCaml releases, but also does not combine
well with the re-strengthening that happens when doing 'include
Exposed_modules' from Coccilib.

This patch inlines the content of Exposed_modules in Coccilib
directly, which gets rid of the double-indirection, and results in
stronger type equalities. The regression test file tests/mdeclp2.cocci
passes after this change.

Long-term I think that it would be nice to have an explicit
coccilib.mli interface with documentation comments, and coccilib.ml
with implementation only. Then you could do either
  coccilib.ml:
    module Foo = Foo
  coccilib.mli:
    module Foo = module type of Foo
or
  coccilib.ml:
    module Foo = struct include Foo end
  coccilib.mli:
    module Foo = module type of Foo
both should work (the first being simpler).